+++ apps/ledger-live-mobile/e2e/specs/speculos/delegate/delegate.ts
@@ -7,12 +7,14 @@ import {
 import { Application } from "../../../page";
 import { CLI } from "../../../utils/cliUtils";
 import { device } from "detox";
+import { Currency } from "@ledgerhq/live-common/e2e/enum/Currency";
+import { getCurrencyManagerApp } from "../../../models/currencies";
 
 export async function runDelegateTest(delegation: Delegate, tmsLink: string) {
   const app = new Application();
 
   $TmsLink(tmsLink);
-  describe(`Delegate flow on  ${delegation.account.currency.name}`, () => {
+  describe(`Delegate`, () => {
     beforeAll(async () => {
       await app.init({
         speculosApp: delegation.account.currency.speculosApp,
@@ -32,25 +34,37 @@ export async function runDelegateTest(delegation: Delegate, tmsLink: string) {
     });
 
     it(`Delegate on ${delegation.account.currency.name}`, async () => {
+      let fees;
       const amountWithCode = delegation.amount + " " + delegation.account.currency.ticker;
-      const currencyId = delegation.account.currency.currencyId;
+      const currencyId =
+        getCurrencyManagerApp(delegation.account.currency.currencyId) ??
+        delegation.account.currency.currencyId;
 
-      await app.accounts.openViaDeeplink();
+      if (delegation.account.currency.name == Currency.INJ.name) {
+        await app.speculos.activateExpertMode();
+      }
+
+      await app.portfolio.goToAccounts(delegation.account.currency.name);
       await app.common.goToAccountByName(delegation.account.accountName);
       await app.account.tapEarn();
 
       await app.stake.dismissDelegationStart(currencyId);
-      await app.stake.setAmount(currencyId, delegation.amount);
-      await app.stake.validateAmount(currencyId);
+      if (delegation.account.currency.name !== Currency.ADA.name) {
+        await app.stake.setAmount(currencyId, delegation.amount);
+        await app.stake.validateAmount(currencyId);
+      } else {
+        await app.stake.verifyFeesVisible(currencyId);
+        fees = await app.stake.getDisplayedFees(currencyId);
+      }
       await app.stake.expectProvider(currencyId, delegation.provider);
       await app.stake.summaryContinue(currencyId);
 
-      await verifyAppValidationStakeInfo(app, delegation, amountWithCode);
+      await verifyAppValidationStakeInfo(app, delegation, amountWithCode, fees);
       await app.speculos.signDelegationTransaction(delegation);
       await device.disableSynchronization();
       await app.common.successViewDetails();
 
-      await verifyStakeOperationDetailsInfo(app, delegation, amountWithCode);
+      await verifyStakeOperationDetailsInfo(app, delegation, amountWithCode, fees);
     });
   });
 }