+++ libs/coin-modules/coin-tron/src/config.ts
@@ -1,5 +1,5 @@
-import { CurrencyConfig, CoinConfig } from "@ledgerhq/coin-framework/config";
-import { MissingCoinConfig } from "@ledgerhq/coin-framework/errors";
+import buildCoinConfig, { CoinConfig, type CurrencyConfig } from "@ledgerhq/coin-framework/config";
+import { CryptoCurrency } from "@ledgerhq/types-cryptoassets";
 
 export type TronConfig = {
   explorer: {
@@ -9,16 +9,9 @@ export type TronConfig = {
 
 export type TronCoinConfig = CurrencyConfig & TronConfig;
 
-let coinConfig: CoinConfig<TronCoinConfig> | undefined;
+const coinConfig: {
+  setCoinConfig: (config: CoinConfig<TronCoinConfig>) => void;
+  getCoinConfig: (currency?: CryptoCurrency) => TronCoinConfig;
+} = buildCoinConfig<TronCoinConfig>();
 
-export const setCoinConfig = (config: CoinConfig<TronCoinConfig>): void => {
-  coinConfig = config;
-};
-
-export const getCoinConfig = (): TronCoinConfig => {
-  if (!coinConfig) {
-    throw new MissingCoinConfig();
-  }
-
-  return coinConfig();
-};
+export default coinConfig;