+++ libs/ledgerjs/packages/types-live/CHANGELOG.md
@@ -1,12 +1,12 @@
 # @ledgerhq/types-live
 
-## 6.60.0-nightly.1
+## 6.60.0
 
 ### Minor Changes
 
-- [#9199](https://github.com/LedgerHQ/ledger-live/pull/9199) [`e4d9a5c`](https://github.com/LedgerHQ/ledger-live/commit/e4d9a5ce6c3e8f2b3829f8f5772e7ba712a4a50c) Thanks [@thesan](https://github.com/thesan)! - Add the `feature_lld_NanoS_upsell_banners` flag
+- [#9241](https://github.com/LedgerHQ/ledger-live/pull/9241) [`0a4413d`](https://github.com/LedgerHQ/ledger-live/commit/0a4413dd5cf50967c27b39bbd7dd9222e6322a60) Thanks [@LucasWerey](https://github.com/LucasWerey)! - Add link param to mevProtection ff
 
-## 6.60.0-nightly.0
+## 6.60.0-next.0
 
 ### Minor Changes